backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lingualive-backend"
version = "0.1.0"
description = "LinguaLive v2 backend"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110,<0.116",
  "uvicorn[standard]>=0.30,<0.32",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "httpx>=0.27,<0.28",
  "python-dotenv>=1.0,<2.0"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
